+++ src/decl/rep-decl.h
@@ -25,14 +25,14 @@ void stop_rep_size_oob(struct r_lazy call);
 static
 r_obj* vec_rep_each_uniform(r_obj* x,
                             int times,
-                            struct r_lazy call,
+                            struct r_lazy error_call,
                             struct vctrs_arg* p_times_arg);
 
 static
 r_obj* vec_rep_each_impl(r_obj* x,
                          r_obj* times,
                          const r_ssize times_size,
-                         struct r_lazy call,
+                         struct r_lazy error_call,
                          struct vctrs_arg* p_times_arg);
 
 static inline